requirements.txt
librosa==0.10.1
praat-parselmouth==0.4.3
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
soundfile==0.12.1